Remove unused file levels.f90 from build list Correct a variety of compile warnings. Placing a single "write" statement prior to call to Input in subroutine netrad prevents the segmentation fault.  The fault comes about due to NaN values in a number of parameters, including variable Time.  No idea why a write would change that...

+++ src/gettbl.f90
@@ -20,7 +20,7 @@ subroutine  GETTBL
 
   OPEN (UNIT = 14, FILE = 'LUT.DAT')
 
-  I = 1 + OMEGA * 2
+  I = INT(1 + OMEGA * 2)
   J = I - 1
   IF ( I .GT. 11 ) I = 11
